move ibc-related code to its own module
