+++ crates/astria-sequencer/src/accounts/component.rs
@@ -24,7 +24,7 @@ pub(crate) struct AccountsComponent;
 impl Component for AccountsComponent {
     type AppState = GenesisState;
 
-    #[instrument(name = "AccountsComponent:init_chain", skip(state))]
+    #[instrument(name = "AccountsComponent::init_chain", skip(state))]
     async fn init_chain<S: StateWriteExt>(mut state: S, app_state: &Self::AppState) -> Result<()> {
         let native_asset = get_native_asset();
         for account in &app_state.accounts {
@@ -33,18 +33,10 @@ impl Component for AccountsComponent {
                 .context("failed writing account balance to state")?;
         }
 
-        state
-            .put_ibc_sudo_address(app_state.ibc_sudo_address)
-            .context("failed to set IBC sudo key")?;
-
-        for address in &app_state.ibc_relayer_addresses {
-            state.put_ibc_relayer_address(address);
-        }
-
         Ok(())
     }
 
-    #[instrument(name = "AccountsComponent:begin_block", skip(_state))]
+    #[instrument(name = "AccountsComponent::begin_block", skip(_state))]
     async fn begin_block<S: StateWriteExt + 'static>(
         _state: &mut Arc<S>,
         _begin_block: &BeginBlock,
@@ -52,7 +44,7 @@ impl Component for AccountsComponent {
         Ok(())
     }
 
-    #[instrument(name = "AccountsComponent:end_block", skip(_state))]
+    #[instrument(name = "AccountsComponent::end_block", skip(_state))]
     async fn end_block<S: StateWriteExt + 'static>(
         _state: &mut Arc<S>,
         _end_block: &EndBlock,
