+++ crates/astria-sequencer/src/accounts/mod.rs
@@ -1,7 +1,4 @@
 pub(crate) mod action;
 pub(crate) mod component;
-pub(crate) mod ibc_relayer_change;
-pub(crate) mod ics20_transfer;
-pub(crate) mod ics20_withdrawal;
 pub(crate) mod query;
 pub(crate) mod state_ext;
