+++ crates/astria-sequencer/src/lib.rs
@@ -5,7 +5,7 @@ pub(crate) mod authority;
 pub(crate) mod component;
 pub mod config;
 pub(crate) mod genesis;
-pub(crate) mod host_interface;
+pub(crate) mod ibc;
 #[cfg(feature = "mint")]
 pub(crate) mod mint;
 pub(crate) mod proposal;
